+++ src/chart.component.jsx
@@ -30,10 +30,11 @@ export class Chart extends Component {
 		}
 	};
 
-	render({ children, data }, { rect }) {
+	render({ children, className, data }, { rect }) {
 		const host = this._ref;
+		const cls = ['plotery', this.props.class || className].filter(x => x);
 		return (
-			<div className="plotery">
+			<div className={cls.join(' ')}>
 				<svg overflow="visible" ref={this._setRef}>
 					<Surface>
 						{host && rect && withProps(children, { data, rect, host })}