+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Added modifier props for `Pointer` and `Wheel` based components.
+- Added class props to `Chart` component.
 
 ### Changed